@@ -1,34 +1,42 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.40'
+    ext.kotlin_version = '1.6.21'
     repositories {
-        maven {
+        /*maven {
             url "https://mirrors.tencent.com/nexus/repository/maven-public/"
         }
-        maven {
-            url "https://maven.oa.com/nexus/content/repositories/jcenter"
-        }
-        maven {
-            url "https://maven.oa.com/nexus/content/groups/androidbuild"
-        }
 
-        maven{
-            url "https://maven.oa.com/nexus/content/repositories/thirdparty-snapshots"
-        }
+        // 内网统一访问maven
         maven {
-            url "https://maven.oa.com/nexus/content/repositories/jitpack.io"
+            url "https://mirrors.tencent.com/repository/maven/tencent_public/"
         }
 
         maven {
-            url "https://maven.oa.com/nexus/content/repositories/google-maven/"
+            url "https://mirrors.tencent.com/repository/maven/tencent_public_snapshots/"
+        }
+
+        maven {
+            url "https://mirrors.tencent.com/repository/maven/thirdparty/"
+        }
+        maven {
+            url "https://mirrors.tencent.com/repository/maven/thirdparty-snapshots/"
+        }
+        maven { url "https://plugins.gradle.org/m2/" }
+
+        mavenCentral()*/
+        maven {
+            url codingArtifactsRepoUrl
+            credentials {
+                username = codingArtifactsGradleUsername
+                password = codingArtifactsGradlePassword
+            }
         }
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:7.1.3'
+        classpath 'com.android.tools.build:gradle:7.2.2'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.kezong:fat-aar:1.2.8'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -37,25 +45,13 @@ buildscript {
 allprojects {
     repositories {
         maven {
-            url "https://mirrors.tencent.com/nexus/repository/maven-public/"
+            url codingArtifactsRepoUrl
+            credentials {
+                username = codingArtifactsGradleUsername
+                password = codingArtifactsGradlePassword
+            }
         }
-        maven {
-            url "https://maven.oa.com/nexus/content/repositories/jcenter"
-        }
-        maven {
-            url "https://maven.oa.com/nexus/content/groups/androidbuild"
-        }
-
-        maven{
-            url "https://maven.oa.com/nexus/content/repositories/thirdparty-snapshots"
-        }
-        maven {
-            url "https://maven.oa.com/nexus/content/repositories/jitpack.io"
-        }
-
-        maven {
-            url "https://maven.oa.com/nexus/content/repositories/google-maven/"
-        }
+        mavenCentral()
     }
 }
 
