@@ -1,11 +1,12 @@
 package com.tencent.joox.sdk
 
 import android.app.Application
+import android.os.Build
 import android.util.Log
 import com.joox.sdklibrary.AuthType
 import com.joox.sdklibrary.SDKInstance
-import com.tencent.ibg.joox.opensdk.openapi.JXApiFactory
-
+import com.tencent.joox.sdk.data.ConstKey
+import com.tencent.joox.sdk.tools.SharedPreferencesTool
 import kotlin.properties.Delegates
 
 class SDKApplication : Application() {
@@ -19,22 +20,15 @@ class SDKApplication : Application() {
         super.onCreate()
         Log.d("SDKApplication", "onCreate called!!!!");
         instance = this
-        val scopeList = ArrayList<String>()
-        scopeList.add(getScopeList())
-
-        JXApiFactory.sDebugLevel = SDKInstance.ENVIR_TYPE
-        val isDebugMode = SDKInstance.ENVIR_TYPE != SDKInstance.PUBLISH
-        val loginType = SharedPreferencesTool.getmInstance(applicationContext).getIntValue(SharedPreferencesTool.LOGIN_TYPE, AuthType.AUTH_WITH_QRCODE)
-
-        // TODO: 2021/7/17  your {AppId} & {App Package Name}
-        SDKInstance.getmInstance().init(instance, null, {ZAppId}, {App Package Name}, loginType, scopeList)
-        SDKInstance.mInstance.setDownFileDir(externalCacheDir.absolutePath + "/testsong/")
-
+        val scopeList = arrayListOf(getScopeList())
+        val appKey = "input app key"
+        val appPkg = "input app package name"
+        SDKInstance.getIns().init(instance, appKey, appPkg, scopeList)
     }
 
     override fun onTerminate() {
         super.onTerminate()
-        SDKInstance.getmInstance().logout();
+        SDKInstance.getIns().logout();
     }
 
     fun saveScopeList(scopeList: String) {
@@ -42,7 +36,8 @@ class SDKApplication : Application() {
     }
 
     fun getScopeList(): String {
-        return SharedPreferencesTool.getmInstance(applicationContext).getStringValue(ConstKey.SCOPE_KEY, "public+user_profile+playmusic_free+search+user_music")
+        return SharedPreferencesTool.getmInstance(applicationContext)
+                .getStringValue(ConstKey.SCOPE_KEY, "public+user_profile+playmusic+search+user_music")
     }
 
 }
